fix: redo custom hook to context
