+++ src/components/theme-switcher/theme-switcher.tsx
@@ -1,16 +1,11 @@
-import { useThemeSwitcher } from '@hooks/UseTheme';
+import { useTheme } from '@hooks/UseTheme';
 import styles from './theme-switcher.module.css';
 
 export function ThemeSwitcher() {
-  const [theme, toggleTheme] = useThemeSwitcher();
+  const { theme, toggleTheme } = useTheme();
 
   return (
-    <button
-      type="button"
-      className={`theme-toggle ${theme === 'dark' ? 'dark' : 'light'}`}
-      onClick={toggleTheme}
-      aria-label="Toggle theme"
-    >
+    <button type="button" className={`theme-toggle ${theme}`} onClick={toggleTheme} aria-label="Toggle theme">
       <span className={styles.thumb}>{theme === 'dark' ? '🌙' : '☀️'}</span>
     </button>
   );
