+++ src/consts/const.ts
@@ -3,3 +3,4 @@ export const LIMIT_POKEMON = 1300;
 export const LAST_TERM = localStorage.getItem('searchTerm');
 export const API = 'https://pokeapi.co/api/v2/pokemon';
 export const PAGE_BACK = -1;
+export const THEME_KEY = 'theme';
