+++ src/context/themeContext.tsx
@@ -1,0 +1,25 @@
+import { THEME_KEY } from '@const/const';
+import { createContext, useEffect, useMemo, useState } from 'react';
+import { Theme, ThemeContextType } from 'src/types/types';
+
+export const ThemeContext = createContext<ThemeContextType | undefined>(undefined);
+
+export function ThemeProvider({ children }: { children: React.ReactNode }) {
+  const [theme, setTheme] = useState<Theme>(() => {
+    const stored = localStorage.getItem(THEME_KEY);
+    return stored === 'dark' ? 'dark' : 'light';
+  });
+
+  useEffect(() => {
+    document.documentElement.setAttribute('data-theme', theme);
+    localStorage.setItem(THEME_KEY, theme);
+  }, [theme]);
+
+  const toggleTheme = () => {
+    setTheme(prev => (prev === 'light' ? 'dark' : 'light'));
+  };
+
+  const themeContextValue = useMemo(() => ({ theme, toggleTheme }), [theme]);
+
+  return <ThemeContext.Provider value={themeContextValue}>{children}</ThemeContext.Provider>;
+}
