+++ src/hooks/UseTheme.tsx
@@ -1,23 +1,9 @@
-import { useEffect, useState } from 'react';
+import { useContext } from 'react';
+import { ThemeContextType } from 'src/types/types';
+import { ThemeContext } from '../context/themeContext';
 
-type Theme = 'light' | 'dark';
-
-const THEME_KEY = 'theme';
-
-export function useThemeSwitcher(): [Theme, () => void] {
-  const [theme, setTheme] = useState<Theme>(() => {
-    const saved = localStorage.getItem(THEME_KEY);
-    return saved === 'dark' ? 'dark' : 'light';
-  });
-
-  useEffect(() => {
-    document.documentElement.setAttribute('data-theme', theme);
-    localStorage.setItem(THEME_KEY, theme);
-  }, [theme]);
-
-  const toggleTheme = () => {
-    setTheme(prev => (prev === 'light' ? 'dark' : 'light'));
-  };
-
-  return [theme, toggleTheme];
+export function useTheme(): ThemeContextType {
+  const context = useContext(ThemeContext);
+  if (!context) throw new Error('useTheme must be wrapped in ThemeProvider');
+  return context;
 }
