+++ src/types/types.ts
@@ -1,0 +1,5 @@
+export type Theme = 'light' | 'dark';
+export type ThemeContextType = {
+  theme: Theme;
+  toggleTheme: () => void;
+};
